Fix bug old config backup code
While moving files, ensure that file names which already exists at the
backup location are not overwritten. If a similarly named file exists at
the backup location use an appropriate numeric suffix to differentiate
the new file.

When creating symlinks, ensure that the target directory heirarchy
exists and if not create it.

Make the equality check of $HOME/.vim with $SCRIPT_DIR more robust.